+++ backend/tournesol/models/__init__.py
@@ -3,3 +3,5 @@ Models for Tournesol app
 """
 
 from .video import *
+from .ratings import *
+from .comparisons import *
