+++ backend/tournesol/models/ratings.py
@@ -1,0 +1,70 @@
+"""
+Models for Tournesol's main functions related to contributor's ratings
+"""
+
+from languages.languages import LANGUAGES
+
+from django.db import models
+
+from core.models import User
+from core.utils.models import WithFeatures
+
+from .video import Video
+
+
+class ContributorRating(models.Model, WithFeatures):
+    """Predictions by individual contributor models."""
+
+    video = models.ForeignKey(
+        Video,
+        on_delete=models.CASCADE,
+        help_text="Video being scored",
+        related_name="contributorvideoratings",
+    )
+    user = models.ForeignKey(
+        User,
+        on_delete=models.CASCADE,
+        help_text="The contributor",
+        related_name="contributorvideoratings",
+    )
+    is_public = models.BooleanField(
+        default=False, null=False, help_text="Should the rating be public?"
+    )
+
+    class Meta:
+        unique_together = ["user", "video"]
+
+    def __str__(self):
+        return "%s on %s" % (self.user, self.video)
+
+
+class ContributorRatingCriteriaScore(models.Model):
+    """Scores per criteria for contributors' algorithmic representatives"""
+
+    contributor_rating = models.ForeignKey(
+        to=ContributorRating,
+        help_text="Foreign key to the contributor's rating",
+        related_name="criteria_scores",
+        on_delete=models.CASCADE,
+    )
+    criteria = models.TextField(
+        max_length=32,
+        help_text="Name of the criteria",
+        db_index=True,
+    )
+    score = models.FloatField(
+        default=0,
+        blank=False,
+        help_text="Score for the given criteria",
+    )
+    uncertainty = models.FloatField(
+        default=0,
+        blank=False,
+        help_text="Uncertainty about the video's score for the given criteria",
+    )
+
+    class Meta:
+        unique_together = ["contributor_rating", "criteria"]
+
+    def __str__(self):
+        return f"{self.contributor_rating}/{self.criteria}/{self.score}"
